domain/src/test/java/com/java/althgriom2022/trees/BstKthLargestTest.java
package com.java.althgriom2022.trees;

import com.java.althgriom.leetcode.entity.TreeNode;
import org.junit.Assert;
import org.junit.Test;

import static org.junit.Assert.*;

/**
 * @Description:
 * @Author: leiline
 * @CreateTime: 2022-03-27
 */
public class BstKthLargestTest {

    @Test
    public void kthLargestTest() {

        TreeNode root = new TreeNode(5);
        TreeNode node1 = new TreeNode(3);
        TreeNode node2 = new TreeNode(6);
        TreeNode node3 = new TreeNode(2);
        TreeNode node4 = new TreeNode(4);
        TreeNode node5 = new TreeNode(1);

        root.left = node1;
        root.right = node2;
        node2.left = node3;
        node2.right = node4;
        node3.left = node5;

        Assert.assertEquals(BstKthLargest.kthLargest(root, 3), 4);
    }
}